% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{csd_population_density_quantiles}
\alias{csd_population_density_quantiles}
\title{Breaks for CSD population density used in map}
\format{
An object of class \code{numeric} of length 6.
}
\usage{
csd_population_density_quantiles
}
\description{
Breaks for CSD population density used in map
}
\keyword{datasets}
