% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{csd_quantiles_text}
\alias{csd_quantiles_text}
\title{Legend for breaks for CSD population density used in map}
\format{
An object of class \code{glue} (inherits from \code{character}) of length 5.
}
\usage{
csd_quantiles_text
}
\description{
Legend for breaks for CSD population density used in map
}
\keyword{datasets}
