% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ridings_population_density_quantiles}
\alias{ridings_population_density_quantiles}
\title{Breaks for riding population density used in map}
\format{
An object of class \code{numeric} of length 6.
}
\usage{
ridings_population_density_quantiles
}
\description{
Breaks for riding population density used in map
}
\keyword{datasets}
