% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{variable_definitions}
\alias{variable_definitions}
\title{Variable definitions}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 38 rows and 1 columns.
}
\usage{
variable_definitions
}
\description{
Variable definitions
}
\keyword{datasets}
