% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vectors}
\alias{vectors}
\title{Vectors used in app}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 640 rows and 9 columns.
}
\usage{
vectors
}
\description{
Vectors used in app
}
\keyword{datasets}
