% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data.R
\name{prepare_data}
\alias{prepare_data}
\title{Prepare data for export/report}
\usage{
prepare_data(geography, regions)
}
\arguments{
\item{geography}{CT or CSD}

\item{regions}{Vector of geo_uids}
}
\description{
Prepare data for export/report
}
